requirements.txt
matplotlib==3.7.1
pandas==1.5.2
lru-dict==1.2.0
scikit-learn==1.2.0
pathlib==1.0.1
codecarbon==2.3.1
